Change encoding order in TLG8

+++ src/tlg8_predictors.h
@@ -10,7 +10,7 @@ namespace tlg::v8::enc
   // インデックスで参照するため、順序は固定する。
   constexpr uint32_t kNumPredictors = 8;
 
-  using predictor_fn = uint8_t (*)(uint8_t, uint8_t, uint8_t, uint8_t);
+  using predictor_fn = int16_t (*)(int16_t, int16_t, int16_t, int16_t);
 
   const std::array<predictor_fn, kNumPredictors> &predictor_table();
 }
